+++ frontend/app/src/components/settings/api-keys/ExchangeKeysForm.vue
@@ -129,8 +129,10 @@
 
 <script lang="ts">
 import { Component, Emit, Prop, Vue } from 'vue-property-decorator';
+import { mapGetters } from 'vuex';
 import ExchangeDisplay from '@/components/display/ExchangeDisplay.vue';
 import BinancePairsSelector from '@/components/helper/BinancePairsSelector.vue';
+import { tradeLocations } from '@/components/history/consts';
 import RevealableInput from '@/components/inputs/RevealableInput.vue';
 import {
   EXCHANGE_BINANCE,
@@ -147,11 +149,15 @@ import { trimOnPaste } from '@/utils/event';
 
 @Component({
   name: 'ExchangeKeysForm',
-  components: { RevealableInput, ExchangeDisplay, BinancePairsSelector }
+  components: { RevealableInput, ExchangeDisplay, BinancePairsSelector },
+  computed: {
+    ...mapGetters('balances', ['exchangeNonce'])
+  }
 })
 export default class ExchangeKeysForm extends Vue {
   readonly krakenAccountTypes = KRAKEN_ACCOUNT_TYPES;
   readonly exchanges = SUPPORTED_EXCHANGES;
+  exchangeNonce!: (exchange: SupportedExchange) => number;
 
   @Prop({ required: true, type: Boolean })
   value!: boolean;
@@ -212,10 +218,21 @@ export default class ExchangeKeysForm extends Vue {
     return exchange === EXCHANGE_BINANCE || exchange === EXCHANGE_BINANCEUS;
   }
 
+  mounted() {
+    if (this.edit) {
+      return;
+    }
+    this.onUpdateExchange({
+      ...this.exchange,
+      name: this.suggestedName(this.exchange.location)
+    });
+  }
+
   onExchangeChange(exchange: SupportedExchange) {
     (this.$refs.form as any).reset();
+    const name = this.suggestedName(exchange);
     this.onUpdateExchange({
-      name: '',
+      name: name,
       newName: null,
       location: exchange,
       apiKey: null,
@@ -224,6 +241,13 @@ export default class ExchangeKeysForm extends Vue {
       krakenAccountType: exchange === EXCHANGE_KRAKEN ? 'starter' : null,
       binanceMarkets: null
     });
+  }
+
+  private suggestedName(exchange: SupportedExchange): string {
+    const location = tradeLocations.find(
+      ({ identifier }) => identifier === exchange
+    );
+    return location ? `${location.name} ${this.exchangeNonce(exchange)}` : '';
   }
 
   onApiKeyPaste(event: ClipboardEvent) {